switch back to a run command in plan, add bundler to load path in dcob binstub
$(hab pkg path yadda/foo) in the run hook would bring in the latest yadda/foo
on a system, not necessarily the yadda/foo a package is built against. So,
that'd be bad to keep using.

This ratchets back to the short run command given in the path with
pkg_svc_run and makes the dcob binstub responsible for adding the
package-local bundle path to the load path and requiring bundler setup to
get the goodness of the equivalent to `bundle exec`.

Signed-off-by: Robb Kidd <robb@thekidds.org>