add a new payload variant for repository events
* Create a hookit method and place it in a GitHub client utility class
* Class method added to Octoclient to make calling hookit read better
* Figure out webhook to add based on incoming request URL
* Add TESTS!
* Use chefstyle and make our world better

Because these webhooks are signed, we can trust that the incoming request
URL for the repository event is the same URL we want to use for the
webhook on pull_request events. This simplifies configuration by assuming
that the incoming URL should the URL used for creating the webhook in
hookit().

Signed-off-by: Robb Kidd <robb@thekidds.org>