move the run hook back into the plan
I created a separate run hook script so that I could set the
production runtime variables for where to fit SSL certs and
to switch the Rack app to production mode. After some discussion,
Thom May and I agreed that those should be concerns for how one
decides to start the bot (e.g. in the systemd or upstart job
definition that starts the hab supervisor).

Updated README with requirement to set production env vars
in whatever is responsible for starting the supervisor.

Signed-off-by: Robb Kidd <robb@thekidds.org>